Fix build failure due to missing fallback iniramfs

+++ build-image.sh
@@ -216,10 +216,18 @@ mkdir -p /usr/var/lib/pacman
 cp -r /var/lib/pacman/local /usr/var/lib/pacman/
 
 # move kernel image and initrd to a defualt location if "linux" is not used
-if [ ${KERNEL_PACKAGE} != 'linux' ] ; then
-	mv /boot/vmlinuz-${KERNEL_PACKAGE} /boot/vmlinuz-linux
-	mv /boot/initramfs-${KERNEL_PACKAGE}.img /boot/initramfs-linux.img
-	mv /boot/initramfs-${KERNEL_PACKAGE}-fallback.img /boot/initramfs-linux-fallback.img
+if [ ${KERNEL_PACKAGE} != 'linux' ]; then
+	if [ -f /boot/vmlinuz-${KERNEL_PACKAGE} ]; then
+		mv /boot/vmlinuz-${KERNEL_PACKAGE} /boot/vmlinuz-linux
+	fi
+
+	if [ -f /boot/initramfs-${KERNEL_PACKAGE}.img ]; then
+		mv /boot/initramfs-${KERNEL_PACKAGE}.img /boot/initramfs-linux.img
+	fi
+
+	if [ -f /boot/initramfs-${KERNEL_PACKAGE}-fallback.img ]; then
+		mv /boot/initramfs-${KERNEL_PACKAGE}-fallback.img /boot/initramfs-linux-fallback.img
+	fi
 fi
 
 # clean up/remove unnecessary files
